+++ src/utils/constants/altTexts.js
@@ -1,0 +1,8 @@
+export const ALT_MARK_TEXT = "Mark task as done";
+export const ALT_EDIT_TEXT = "Edit task";
+export const ALT_DELETE_TEXT = "Delete task";
+export const ALT_LOGO_TEXT = "Logo";
+export const ALT_SEARCH_TEXT = "Search";
+export const ALT_NAV_LOGO_TEXT = "Logo";
+export const ALT_ADD_TEXT = "Add Task Button";
+export const ALT_HIDE_BUTTON_TEXT = "Hide Create Button";
