+++ src/utils/constants/classNames.js
@@ -1,0 +1,2 @@
+export const CLASS_BUTTON_CREATE = "content__button--create";
+export const CLASS_TASK_INPUT = "task-input";
