+++ src/utils/constants/constants.js
@@ -1,17 +1,3 @@
-export const ALT_MARK_TEXT = "Mark task as done";
-export const ALT_EDIT_TEXT = "Edit task";
-export const ALT_DELETE_TEXT = "Delete task";
-export const ALT_LOGO_TEXT = "Logo";
-export const ALT_SEARCH_TEXT = "Search";
-export const ALT_NAV_LOGO_TEXT = "Logo";
-export const ALT_ADD_TEXT = "Add Task Button";
-export const ALT_HIDE_BUTTON_TEXT = "Hide Create Button";
-
 export const TEXT_CREATE = "create";
 export const TEXT_ENTER = "Enter";
 export const TEXT_ADD = "Add";
-
-export const CLASS_BUTTON_CREATE = "content__button--create";
-export const CLASS_TASK_INPUT = "task-input";
-
-export const TYPE_TEXT = "TYPE_TEXT";
