+++ src/utils/constants/typeTexts.js
@@ -1,0 +1,1 @@
+export const TYPE_TEXT = "TYPE_TEXT";
